api-users/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]
passlib[bcrypt]==1.7.4
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.12
email_validator==2.2.0
bcrypt==4.0.1
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.27.0
